Add CephSecrets to OpenStackDataPlaneNode
If CephSecrets list is non-empty then look for each secret
and mount it in the OpenStack Ansible EE pod so that the
edpm_ceph_client_files Ansible role can copy those files
to any OpenStackDataPlaneNode. This results in data plane
nodes having Ceph configuration files (like ceph.conf or
a cephx key) so that OpenStack services on those nodes can
be clients of a Ceph cluster.

Closes: OSP-22956
Signed-off-by: John Fulton <fulton@redhat.com>